requirements.txt
## Create the "fireworker" python virtual environment and select it in this directory:
#   pyenv install 3.8.0
#
#   pyenv local 3.8.0
#   pip install --upgrade pip setuptools virtualenv virtualenvwrapper virtualenv-clone wheel
#
#   pyenv virtualenv fireworker
#   pyenv local fireworker
#   pip install --upgrade pip setuptools virtualenv virtualenvwrapper virtualenv-clone wheel
#   pip install -r requirements.txt
#   pyenv rehash

FireWorks==1.9.5
docker==4.1.0
requests==2.22.0
ruamel.yaml==0.16.5
subprocess32==3.5.4

## The following requirements were added by pip freeze:
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-paginate==0.5.5
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
monty==3.0.2
pbr==5.4.4
pymongo==3.10.0
python-dateutil==2.8.1
six==1.13.0
stevedore==1.31.0
tabulate==0.8.6
tqdm==4.41.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
websocket-client==0.57.0
Werkzeug==0.16.0
